Refactor tests to seperate files
