+++ src/test/mod.rs
@@ -1,0 +1,7 @@
+mod interpret;
+mod args;
+mod ast;
+mod lexer;
+mod main;
+mod parser;
+mod token;
